requirements.txt
dataclasses_json~=0.6.4
llama_index~=0.10.18
python-dotenv~=1.0.1
tqdm~=4.66.2
typing_extensions~=4.10.0
llama-index-graph-stores-neo4j
llama-index-vector-stores-neo4jvector
llama-index-llms-ollama
chainlit
llama-index-embeddings-huggingface

--index-url https://download.pytorch.org/whl/cu118
torch
pcst_fast
pyg_lib
torch_geometric